requirement.txt
selenium
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
requests